+++ lib/ui/screens/movies/home.dart
@@ -1,104 +1,29 @@
-import 'package:cinefy/ui/providers/providers.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../domain/entities/movie_entity.dart';
+import '../../views/views.dart';
 import '../../widgets/widgets.dart';
 
 class HomeScreen extends StatelessWidget {
   static const String name = 'home-screen';
+  final int viewIndex;
 
-  const HomeScreen({super.key});
+  const HomeScreen({super.key, required this.viewIndex});
+
+  final viewRoutes = const <Widget>[
+    InicioView(),
+    FavoritesView(),
+    CategoriesView(),
+  ];
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: _HomeView(),
-      bottomNavigationBar: CustomBottomNavigator(),
-    );
-  }
-}
-
-class _HomeView extends ConsumerStatefulWidget {
-  const _HomeView();
-
-  @override
-  _HomeViewState createState() => _HomeViewState();
-}
-
-class _HomeViewState extends ConsumerState<_HomeView> {
-  @override
-  void initState() {
-    super.initState();
-    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-    ref.read(popularMoviesProvider.notifier).loadNextPage();
-    ref.read(upcomingMoviesProvider.notifier).loadNextPage();
-    ref.read(topRatedMoviesProvider.notifier).loadNextPage();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (ref.watch(firstLoadingProvider)) return const SplashLoader();
-
-    final List<Movie> nowPlayingMovies = ref.watch(nowPlayingMoviesProvider);
-    final List<Movie> popularMovies = ref.watch(popularMoviesProvider);
-    final List<Movie> upcomingMovies = ref.watch(upcomingMoviesProvider);
-    final List<Movie> topRatedMovies = ref.watch(topRatedMoviesProvider);
-    final List<Movie> slideShowMovies = ref.watch(moviesSlideshowProvider);
-
-    return CustomScrollView(
-      slivers: [
-        const SliverAppBar(
-          leadingWidth: 0,
-          floating: true,
-          flexibleSpace: FlexibleSpaceBar(
-            title: CustomAppbar(),
-          ),
-        ),
-        SliverList(
-          delegate: SliverChildBuilderDelegate(
-            (context, index) {
-              return Column(
-                children: [
-                  // const CustomAppbar(),
-                  MoviesSlideshow(movies: slideShowMovies),
-                  MovieHorizontalList(
-                    movies: nowPlayingMovies,
-                    title: 'En cines',
-                    subTitle: 'Viernes 26',
-                    loadNextPage: () => ref
-                        .read(nowPlayingMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  MovieHorizontalList(
-                    movies: popularMovies,
-                    title: 'Populares',
-                    subTitle: 'Top calientes',
-                    loadNextPage: () =>
-                        ref.read(popularMoviesProvider.notifier).loadNextPage(),
-                  ),
-                  MovieHorizontalList(
-                    movies: topRatedMovies,
-                    title: 'Mejor calificadas',
-                    loadNextPage: () => ref
-                        .read(topRatedMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  MovieHorizontalList(
-                    movies: upcomingMovies,
-                    title: 'Próximamente',
-                    subTitle: 'EL mes que viene',
-                    loadNextPage: () => ref
-                        .read(upcomingMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  const SizedBox(height: 40),
-                ],
-              );
-            },
-            childCount: 1,
-          ),
-        ),
-      ],
+    return Scaffold(
+      body: IndexedStack(
+        index: viewIndex,
+        children: viewRoutes,
+      ),
+      bottomNavigationBar: CustomBottomNavigator(
+        currentIndex: viewIndex,
+      ),
     );
   }
 }
