+++ lib/ui/views/movies/categories_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CategoriesView extends StatelessWidget {
+  const CategoriesView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
