+++ lib/ui/views/movies/favorites_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class FavoritesView extends StatelessWidget {
+  const FavoritesView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
