+++ lib/ui/views/views.dart
@@ -1,0 +1,5 @@
+export 'package:cinefy/ui/views/movies/categories_view.dart';
+
+export 'package:cinefy/ui/views/movies/favorites_view.dart';
+
+export 'package:cinefy/ui/views/movies/inicio_view.dart';
