+++ lib/ui/widgets/share/custom_bottom_navigation.dart
@@ -1,11 +1,31 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigator extends StatelessWidget {
-  const CustomBottomNavigator({Key? key}) : super(key: key);
+  final int currentIndex;
+  const CustomBottomNavigator({Key? key, required this.currentIndex})
+      : super(key: key);
+
+  void onItemTap(BuildContext context, int index) {
+    switch (index) {
+      case 0:
+        context.go('/home/0');
+        break;
+      case 1:
+        context.go('/home/1');
+        break;
+      case 2:
+        context.go('/home/2');
+        break;
+      default:
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
+      currentIndex: currentIndex,
+      onTap: (value) => onItemTap(context, value),
       elevation: 0,
       items: const [
         BottomNavigationBarItem(
